requirements.txt
gradio==4.14.0
opencv-python
scikit-learn
numpy
joblib
pillow
pydub
